+++ MLEM.Ui/Style/StyleProp.cs
@@ -9,35 +9,30 @@ namespace MLEM.Ui.Style {
     /// Note that <c>T</c> implicitly converts to <c>StyleProp{T}</c> and vice versa.
     /// </summary>
     /// <typeparam name="T">The type of style setting that this property stores</typeparam>
-    public struct StyleProp<T> : IEquatable<StyleProp<T>> {
+    public readonly struct StyleProp<T> : IEquatable<StyleProp<T>> {
+
+        /// <summary>
+        /// The empty style property, with no <see cref="Value"/> and a priority of 0.
+        /// </summary>
+        public static StyleProp<T> None => default;
 
         /// <summary>
         /// The currently applied style
         /// </summary>
-        public T Value { get; private set; }
-        private byte priority;
+        public readonly T Value;
+        private readonly byte priority;
 
         /// <summary>
         /// Creates a new style property with the given custom style and a priority of <see cref="byte.MaxValue"/>.
+        /// To create a style property with a lower priority, use <see cref="OrStyle(T,byte)"/> on an existing priority, or use <see cref="None"/>.
         /// </summary>
         /// <param name="value">The custom style to apply</param>
-        public StyleProp(T value) {
-            this.priority = byte.MaxValue;
-            this.Value = value;
+        public StyleProp(T value) : this(value, byte.MaxValue) {
         }
 
-        /// <summary>
-        /// Sets this style property's value and marks it as being set by a <see cref="UiStyle"/> if it doesn't have a custom value yet.
-        /// This allows this property to be overridden by custom style settings using <see cref="StyleProp{T}(T)"/> or a higher <paramref name="priority"/>.
-        /// </summary>
-        /// <param name="value">The style to apply</param>
-        /// <param name="priority">The priority that this style value has. Higher priority style values will override lower priority style values.</param>
-        ///<seealso cref="CopyFromStyle"/>
-        public void SetFromStyle(T value, byte priority = 0) {
-            if (priority >= this.priority) {
-                this.Value = value;
-                this.priority = priority;
-            }
+        private StyleProp(T value, byte priority) {
+            this.Value = value;
+            this.priority = priority;
         }
 
         /// <summary>
@@ -46,12 +41,19 @@ namespace MLEM.Ui.Style {
         /// </summary>
         /// <param name="value">The style to apply</param>
         /// <param name="priority">The priority that the style value has. Higher priority style values will override lower priority style values.</param>
-        /// <returns>The new style</returns>
-        /// <seealso cref="SetFromStyle"/>
-        public StyleProp<T> CopyFromStyle(T value, byte priority = 0) {
-            var ret = this;
-            ret.SetFromStyle(value, priority);
-            return ret;
+        /// <returns>The style with the higher priority</returns>
+        public StyleProp<T> OrStyle(T value, byte priority = 0) {
+            return this.OrStyle(new StyleProp<T>(value, priority));
+        }
+
+        /// <summary>
+        /// Chooses and returns the style property with the higher priority, out of this value and <paramref name="other"/>.
+        /// This allows this property to be overridden by custom style settings using <see cref="StyleProp{T}(T)"/> or a higher priority.
+        /// </summary>
+        /// <param name="other">The style property to compare with</param>
+        /// <returns>The style property with the higher priority</returns>
+        public StyleProp<T> OrStyle(StyleProp<T> other) {
+            return other.priority >= this.priority ? other : this;
         }
 
         /// <summary>
@@ -64,7 +66,7 @@ namespace MLEM.Ui.Style {
         }
 
         /// <summary>
-        /// Returns whether this style property has a value assigned to it using <see cref="SetFromStyle"/> or <see cref="StyleProp{T}(T)"/>.
+        /// Returns whether this style property has a value assigned to it using <see cref="OrStyle(StyleProp{T})"/> or <see cref="StyleProp{T}(T)"/>.
         /// </summary>
         /// <returns>Whether this style property has a value</returns>
         public bool HasValue() {